chore: remove verdaccio, lerna in favor of commonjs in layers
