--- .npmrc
@@ -1,3 +1,0 @@
-# needed for publishing via yarn when verdaccio publishing is set to "all" 
-# https://github.com/verdaccio/verdaccio/issues/212#issuecomment-308578500
-//localhost:4873/:_authToken="superfakeauthtoken"
