+++ amplify/backend/function/discordlayer/opt/commands/static-commands.json
