+++ amplify/backend/function/discordlayer/parameters.json
@@ -2,5 +2,5 @@
   "runtimes": [
     "nodejs14.x"
   ],
-  "description": "Updated layer version 2021-09-21T22:24:56.840Z"
+  "description": "Updated layer version 2021-09-21T22:53:54.222Z"
 }
