--- lerna.json
@@ -1,5 +1,0 @@
-{
-  "packages": ["packages/*"],
-  "npmClient": "yarn",
-  "version": "independent"
-}
