+++ package.json
@@ -3,25 +3,16 @@
   "version": "0.1.0",
   "license": "MIT",
   "private": true,
-  "workspaces": {
-    "packages": [
-      "packages/*"
-    ]
-  },
+  "workspaces": [
+    "packages/*"
+  ],
   "scripts": {
-    "start": "echo 'no start command needed'",
-    "registry:start": "verdaccio --config ./verdaccio.yaml",
-    "registry:publish": "lerna publish from-package --no-git-tag-version --no-push --canary --preid next --pre-dist-tag next --registry http://localhost:4873 --yes",
-    "registry": "run-p registry:*"
+    "start": "echo 'no start command needed'"
   },
   "devDependencies": {
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
-    "lerna": "^4.0.0",
-    "nodemon": "^2.0.12",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.4.0",
-    "verdaccio": "^5.1.5"
+    "prettier": "^2.4.0"
   },
   "prettier": {
     "tabWidth": 2,
