--- packages/bank/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "@amplify-discord-bots/bank",
-  "version": "0.1.1-next.2",
-  "license": "MIT",
-  "type": "module",
-  "files": [
-    "lib"
-  ],
-  "exports": {
-    ".": {
-      "import": "./bank.js",
-      "require": "./lib/bank.cjs"
-    }
-  },
-  "scripts": {
-    "build": "esbuild ./*.js --outdir=lib --platform=node --format=cjs --target=node14 --out-extension:.js=.cjs",
-    "prepublish": "yarn build"
-  },
-  "dependencies": {},
-  "devDependencies": {
-    "esbuild": "^0.12.28"
-  }
-}
