--- packages/command-giverole/.gitignore
@@ -1,1 +1,0 @@
-lib
