--- packages/command-giverole/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@amplify-discord-bots/command-giverole",
-  "version": "0.1.1-next.1",
-  "license": "MIT",
-  "type": "module",
-  "files": [
-    "lib"
-  ],
-  "exports": {
-    ".": {
-      "import": "./giverole.js",
-      "require": "./lib/giverole.cjs"
-    }
-  },
-  "scripts": {
-    "build": "esbuild ./giverole.js --outdir=lib --platform=node --format=cjs --target=node14 --out-extension:.js=.cjs",
-    "prepublish": "yarn build"
-  },
-  "dependencies": {
-    "@amplify-discord-bots/discord": "*"
-  },
-  "devDependencies": {
-    "esbuild": "^0.12.28"
-  }
-}
