--- packages/discord/.gitignore
@@ -1,1 +1,0 @@
-lib
