--- packages/discord/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@amplify-discord-bots/discord",
-  "version": "0.1.1-next.2",
-  "type": "module",
-  "license": "MIT",
-  "files": [
-    "lib"
-  ],
-  "exports": {
-    ".": {
-      "import": "./discord.js",
-      "require": "./lib/discord.cjs"
-    }
-  },
-  "scripts": {
-    "build": "esbuild ./discord.js --outdir=lib --platform=node --format=cjs --target=node14 --out-extension:.js=.cjs",
-    "prepublish": "yarn build"
-  },
-  "dependencies": {
-    "@aws-sdk/client-ssm": "^3.21.0",
-    "node-fetch": "^2.6.1"
-  },
-  "devDependencies": {
-    "esbuild": "^0.12.28"
-  }
-}
