--- verdaccio.yaml
@@ -1,11 +1,0 @@
-storage: ./.verdaccio
-uplinks:
-  npmjs:
-    url: https://registry.npmjs.org/
-packages:
-  '@amplify-discord-bots/*':
-    access: $all
-    publish: $all
-  '**':
-    proxy: npmjs
-logs: { type: stdout, format: pretty, level: http }
